Uploading Mini-Assignment 5
Added three variables: $heir, which tracks the senior child in the line of succession as a text string; $year, which tracks the date as a number value ; and $protrefo which tracks whether or not players have split from the Catholic Church as a boolean value. Added more connections between passages. Removed the passage "Elizabeth-Denmark" and the dependent ending passage associated with it. Included two new options to replace. Expanded/changed some text within passages, and fixed grammatical/spelling errors.